x64: Migrate div instructions to the new assembler (#10820)
* x64: Migrate div instructions to the new assembler

This mostly required adding a new kind of operand representing a
`TrapCode` and plumbing around a few bits and bobs to ensure that this
compile-time-only-abstraction does not need to be accounted for in all
the encoding bits.

* Fix doc tests

* Shrink the size of `Amode`

Reduces the size of `KnownOffset` so `Option<KnownOffset>` doesn't have
such a high alignment or size, greatly shrinking the size of the
instruction and prevents the changes to `Div` from increasing the size
of the overall `Inst` enum. In fact it now shrinks!

* Don't model a trap code as an operand

* Render trap information in `Display`

+++ cranelift/codegen/src/isa/x64/inst/external.rs
@@ -397,8 +397,8 @@ impl Into<asm::Amode<Gpr>> for Amode {
 /// `KnownOffsetTable`.
 #[expect(missing_docs, reason = "self-describing keys")]
 pub mod offsets {
-    pub const KEY_INCOMING_ARG: usize = 0;
-    pub const KEY_SLOT_OFFSET: usize = 1;
+    pub const KEY_INCOMING_ARG: u8 = 0;
+    pub const KEY_SLOT_OFFSET: u8 = 1;
 }
 
 impl asm::CodeSink for MachBuffer<Inst> {
